requirements.txt
pyserial>=3.5
pandas>=2.0.0
matplotlib>=3.7.0
customtkinter>=5.2.0
openpyxl>=3.1.0
numpy>=1.24.0
pillow>=9.0.0
